backend/requirements.txt
django
psycopg2
uwsgi
djangorestframework
markdown
pyyaml
django-filter
django-cors-headers